Move interface to typings file. Remove wrong declaration on vtex.styleguide.d.ts. Update CHANGELOG.
